fix: make security fixtures PSR-4 compliant
Composer skips the shared fixture file during optimized autoload generation because its classes do not match their PSR-4 paths. Store each controller under its namespace path and let the test autoloader resolve them directly.
